Fixed birthday example
Did 2 things to fix the bug in the birthday example.
(1) Added a sanitizer step in the cellvalue computation that
we need to extract from the model.
(2) Fixed the test for whether a cell value has an error to reflect Plato upgrade.
If neg_FIELD_b returns an object of type Exp, then there is an error;
otherwise, there is no error.