Added automatic extraction of type sanitizer
In addition to automatically extracting the validation code and replicating
it on the client, we can now also automatically extract the type casting
that Rails does automatically when data from a form.  We bundle that type casting
into a new routine on the client called sanitize, which is now used in defining
cellvalue.