Added syntax checker for Plato logic.
To use, call check_plato(constraint-string) from the BirthdatesHelper module.
The first check is a visual one: that reading the string representation of
the constraints doesn't lose anything (which can happen b/c of mismatched parentheses).
So the first thing that happens is the string constraints as well as the result
of reading those string constraints are printed to the screen for visual
comparison.

The second thing is that the constraints are transformed into a canonical
representation--the same representation that pl-fhl-to-js uses.

The third thing is that the canonical representation is analyzed for
a number of syntactic problems.  If problems are found, the canonical
representation is printed along with a list of the problems.

In addition to check_plato, I did some minor reorganizing of the rsbc folder---
moving the notes to a subfolder called NOTES.

Also I updated Plato's version to the latest.

Finally, I changed FebValidator to inherit from EachValidator instead of Validator,
which makes the client-side code work.  Not sure if this is the right thing, but
a helpful workaround at the least.